refactor(settings): extract memory page label/format helpers to own module
The 11 React-free label/format/filter helpers (memoryOriginLabel,
memoryStatusLabel, memoryStatusTone, filterLocalMemoryEntries, the backup
summarizers, etc.) lived at the tail of memory-settings-page.tsx. Move them
to memory-settings-labels.ts and import them back.

memory-settings-labels.ts is registered in SETTINGS_SOURCE_REPO_PATHS right
after memory-settings-page.tsx so the contract source mirror still contains
the same code and the MemoryEntryList .. filterLocalMemoryEntries block
boundary still resolves (now across the two files).
